+++ src/any_value/mod.rs
@@ -1,9 +1,7 @@
-use std::any::{Any, TypeId};
-use std::{mem, ptr};
-use std::marker::PhantomData;
-use std::mem::{ManuallyDrop, MaybeUninit, size_of};
-use std::ops::DerefMut;
-use std::ptr::{drop_in_place, NonNull, null_mut};
+use std::any::{TypeId};
+use std::{ptr};
+use std::mem::{MaybeUninit};
+use std::ptr::{NonNull};
 
 pub(crate) mod temp;
 mod wrapper;