+++ src/any_vec_typed.rs
@@ -1,7 +1,4 @@
-use std::mem::size_of;
 use std::marker::PhantomData;
-use std::mem::MaybeUninit;
-use std::ptr;
 use std::ptr::NonNull;
 use crate::{AnyVec};
 use crate::any_value::{AnyValueTemp, AnyValue, AnyValueWrapper};