+++ src/ops/remove.rs
@@ -1,10 +1,7 @@
-use std::any::{Any, TypeId};
-use std::mem::size_of;
 use std::marker::PhantomData;
-use std::mem::{forget, MaybeUninit};
 use std::ptr;
 use std::ptr::NonNull;
-use crate::{AnyVec, copy_bytes_nonoverlapping, Unknown};
+use crate::{AnyVec, Unknown};
 use crate::any_value::temp::Impl;
 
 /// Lazily `remove` element on consumption/drop.