+++ tests/any_vec.rs
@@ -5,6 +5,7 @@ use itertools::{assert_equal};
 use any_vec::{AnyVec};
 use any_vec::any_value::{AnyValueRaw, AnyValueWrapper};
 
+#[allow(dead_code)]
 unsafe fn any_as_u8_slice<T: Sized>(p: &T) -> &[u8] {
     std::slice::from_raw_parts(
         (p as *const T) as *const u8,