+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.5.0
+### Changed
+- Major type erased API change. All `AnyVec` operations now safe.
+
+### Added
+- Introduced `any_value` family, which provide type erased safe value operations.
+
 ## 0.4.0
 ### Added
 - `AnyVec::remove`